Zoe-2.0/code/python/requirements.txt
groq
mss
Pillow
